# Speed up random access time of PyAVReaderIndexed.