Previously `PyAVReaderIndexed ` decoded every packet to seek to a specific packet. This PR tries to speed this up by instead indexing the timestamps of each packet and seeking to it using `container.seek()`.

This speeds up random access times by an order of magnitude. Sequential access time remains unchanged.

## Test Code
```python
test_reader = PyAVReaderIndexed(...)
inds = np.random.randint(low=0, high=len(test_reader), size=3)
print(test_reader)

def test():
    for i in inds:
        print(i)
        test_reader.get_frame(i)
    
%time test()
```

## Before
```
<Frames>
Source: /local1/kpar/spatialaudiogen/data/preproc-hr/cA88rNC2uuM-video.mp4
Length: 9061 frames
Frame Shape: (1080, 1920, 3)

673
7158
5975
CPU times: user 41.1 s, sys: 7.8 ms, total: 41.1 s
Wall time: 41.1 s
```

## After
```
<Frames>
Source: /local1/kpar/spatialaudiogen/data/preproc-hr/cA88rNC2uuM-video.mp4
Length: 9061 frames
Frame Shape: (1080, 1920, 3)

673
7158
5975
CPU times: user 1.78 s, sys: 8.01 ms, total: 1.79 s
Wall time: 1.78 s
```