@keunhong 2020-04-29
Hi, can anyone take a look at this? I've been using this for the past 2 weeks in my machine learning workflow and I haven't run into any issues yet.

You can even get more gains by re-encoding videos with more frequent keyframes since less frames have to be decoded after seeking.