@sofroniewn 2020-05-01
@keunhong i'm not a PIMS maintainer, but am very excited about this speed-up. Not sure if you've seen my post here https://github.com/dask/dask-image/issues/134 asking about such things for usage with dask, but I'm curious if you'd expect to see speed-ups using `dask-image` `imread` which uses PIMs under the hood. I can give it a try and report back too if others are interested.