@nkeim 2020-05-04
Thanks for submitting this great PR, @keunhong ! And thanks for the review, @caspervdw . I think this should go into v0.5 since it is such a major improvement. The logical alternative would be to wait until v0.6, and I'm scared of that 😉.

@sofroniewn Since this is for random access, I imagine this would help tremendously with any kind of distributed loading of images from PyAV. I hope I am right!